requirements/base.txt
# 基础依赖包
# Web框架
dash>=3.1.1,<4.0.0
flask>=2.3.0,<3.0.0
gunicorn>=21.2.0

# UI组件库
feffery-antd-components>=0.4.0,<0.5.0
feffery-utils-components>=0.3.0,<0.4.0
feffery-dash-utils>=0.2.6

# 数据库ORM
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0          # MySQL

# 缓存
redis>=5.0.0
flask-caching>=2.1.0

# 认证和安全
flask-login>=0.6.0
flask-principal>=0.4.0
pyjwt>=2.8.0
werkzeug>=2.3.0
cryptography>=41.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 工具库
python-dotenv>=1.0.0
click>=8.1.0
user-agents>=2.2.0
yarl>=1.9.0

# 日志和监控
structlog>=23.1.0
sentry-sdk>=1.32.0

# 时间处理
python-dateutil>=2.8.0
pytz>=2023.3